Tag 3.0.2 release.

+++ src/common/filefn.cpp
@@ -37,7 +37,6 @@
 #include "wx/filename.h"
 #include "wx/dir.h"
 
-#include "wx/scopedptr.h"
 #include "wx/tokenzr.h"
 
 // there are just too many of those...
@@ -49,7 +48,11 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include <errno.h>
+#if !wxONLY_WATCOM_EARLIER_THAN(1,4)
+    #if !(defined(_MSC_VER) && (_MSC_VER > 800))
+        #include <errno.h>
+    #endif
+#endif
 
 #if defined(__WXMAC__)
     #include  "wx/osx/private.h"  // includes mac headers
@@ -58,6 +61,7 @@
 #ifdef __WINDOWS__
     #include "wx/msw/private.h"
     #include "wx/msw/missing.h"
+    #include "wx/msw/mslu.h"
 
     // sys/cygwin.h is needed for cygwin_conv_to_full_win32_path()
     // and for cygwin_conv_path()
@@ -100,6 +104,14 @@
 
 #if WXWIN_COMPATIBILITY_2_8
 static wxChar wxFileFunctionsBuffer[4*_MAXPATHLEN];
+#endif
+
+#if defined(__VISAGECPP__) && __IBMCPP__ >= 400
+//
+// VisualAge C++ V4.0 cannot have any external linkage const decs
+// in headers included by more than one primary source
+//
+const int wxInvalidOffset = -1;
 #endif
 
 // ============================================================================
@@ -190,7 +202,7 @@ void wxPathList::AddEnvList (const wxString& WXUNUSED_IN_WINCE(envVariable))
     // "C:\Program" and "Files"; this is true for both Windows and Unix.
 
     static const wxChar PATH_TOKS[] =
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
         wxT(";"); // Don't separate with colon in DOS (used for drive)
 #else
         wxT(":;");
@@ -213,6 +225,13 @@ bool wxPathList::EnsureFileAccessible (const wxString& path)
 {
     return Add(wxPathOnly(path));
 }
+
+#if WXWIN_COMPATIBILITY_2_6
+bool wxPathList::Member (const wxString& path) const
+{
+    return Index(path) != wxNOT_FOUND;
+}
+#endif
 
 wxString wxPathList::FindValidPath (const wxString& file) const
 {
@@ -303,7 +322,7 @@ wxIsAbsolutePath (const wxString& filename)
         if ((filename[0] == wxT('[') && filename[1] != wxT('.')))
             return true;
 #endif
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
         // MSDOS like
         if (filename[0] == wxT('\\') || (wxIsalpha (filename[0]) && filename[1] == wxT(':')))
             return true;
@@ -379,7 +398,7 @@ static CharType *wxDoRealPath (CharType *path)
                         path[0] = SEP;
                         path[1] = wxT('\0');
                       }
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
                     /* Check that path[2] is NULL! */
                     else if (path[1] == wxT(':') && !path[2])
                       {
@@ -509,6 +528,23 @@ static CharType *wxDoExpandPath(CharType *buf, const wxString& name)
 #endif
 
     /* Expand inline environment variables */
+#ifdef __VISAGECPP__
+    while (*d)
+    {
+      *d++ = *s;
+      if(*s == wxT('\\'))
+      {
+        *(d - 1) = *++s;
+        if (*d)
+        {
+          s++;
+          continue;
+        }
+        else
+           break;
+      }
+      else
+#else
     while ((*d++ = *s) != 0) {
 #  ifndef __WINDOWS__
         if (*s == wxT('\\')) {
@@ -519,6 +555,7 @@ static CharType *wxDoExpandPath(CharType *buf, const wxString& name)
                 break;
         } else
 #  endif
+#endif
             // No env variables on WinCE
 #ifndef __WXWINCE__
 #ifdef __WINDOWS__
@@ -732,7 +769,7 @@ wxPathOnly (wxChar *path)
             i --;
         }
 
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
         // Try Drive specifier
         if (wxIsalpha (buf[0]) && buf[1] == wxT(':'))
         {
@@ -784,7 +821,7 @@ wxString wxPathOnly (const wxString& path)
             i --;
         }
 
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
         // Try Drive specifier
         if (wxIsalpha (buf[0]) && buf[1] == wxT(':'))
         {
@@ -902,14 +939,14 @@ void wxDos2UnixFilename(wchar_t *s) { wxDoDos2UnixFilename(s); }
 
 template<typename T>
 static void
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
 wxDoUnix2DosFilename(T *s)
 #else
 wxDoUnix2DosFilename(T *WXUNUSED(s) )
 #endif
 {
 // Yes, I really mean this to happen under DOS only! JACS
-#if defined(__WINDOWS__)
+#if defined(__WINDOWS__) || defined(__OS2__)
   if (s)
     while (*s)
       {
@@ -964,7 +1001,7 @@ wxConcatFiles (const wxString& file1, const wxString& file2, const wxString& fil
 }
 
 // helper of generic implementation of wxCopyFile()
-#if !defined(__WIN32__) && wxUSE_FILE
+#if !(defined(__WIN32__) || defined(__OS2__)) && wxUSE_FILE
 
 static bool
 wxDoCopyFile(wxFile& fileIn,
@@ -1023,6 +1060,9 @@ wxCopyFile (const wxString& file1, const wxString& file2, bool overwrite)
 
         return false;
     }
+#elif defined(__OS2__)
+    if ( ::DosCopy(file1.c_str(), file2.c_str(), overwrite ? DCPY_EXISTING : 0) != 0 )
+        return false;
 #elif wxUSE_FILE // !Win32
 
     wxStructStat fbuf;
@@ -1052,7 +1092,7 @@ wxCopyFile (const wxString& file1, const wxString& file2, bool overwrite)
 
     wxDoCopyFile(fileIn, fbuf, file2, overwrite);
 
-#if defined(__WXMAC__)
+#if defined(__WXMAC__) || defined(__WXCOCOA__)
     // copy the resource fork of the file too if it's present
     wxString pathRsrcOut;
     wxFile fileRsrcIn;
@@ -1092,14 +1132,18 @@ wxCopyFile (const wxString& file1, const wxString& file2, bool overwrite)
         if ( !wxDoCopyFile(fileRsrcIn, fbuf, pathRsrcOut, overwrite) )
             return false;
     }
-#endif // wxMac
+#endif // wxMac || wxCocoa
 
+#if !defined(__VISAGECPP__) && !defined(__WXMAC__) || defined(__UNIX__)
+    // no chmod in VA.  Should be some permission API for HPFS386 partitions
+    // however
     if ( chmod(file2.fn_str(), fbuf.st_mode) != 0 )
     {
         wxLogSysError(_("Impossible to set permissions for the file '%s'"),
                       file2.c_str());
         return false;
     }
+#endif // OS/2 || Mac
 
 #else // !Win32 && ! wxUSE_FILE
 
@@ -1148,6 +1192,8 @@ bool wxRemoveFile(const wxString& file)
 {
 #if defined(__VISUALC__) \
  || defined(__BORLANDC__) \
+ || defined(__WATCOMC__) \
+ || defined(__DMC__) \
  || defined(__GNUWIN32__)
     int res = wxRemove(file);
 #elif defined(__WXMAC__)
@@ -1169,7 +1215,7 @@ bool wxMkdir(const wxString& dir, int perm)
 
     // assume mkdir() has 2 args on non Windows-OS/2 platforms and on Windows too
     // for the GNU compiler
-#elif (!(defined(__WINDOWS__) || defined(__DOS__))) || \
+#elif (!(defined(__WINDOWS__) || defined(__OS2__) || defined(__DOS__))) || \
       (defined(__GNUWIN32__) && !defined(__MINGW32__)) ||                \
       defined(__WINE__) || defined(__WXMICROWIN__)
     const wxChar *dirname = dir.c_str();
@@ -1179,9 +1225,15 @@ bool wxMkdir(const wxString& dir, int perm)
   #else
     if ( mkdir(wxFNCONV(dirname), perm) != 0 )
   #endif
+#elif defined(__OS2__)
+    wxUnusedVar(perm);
+    if (::DosCreateDir(dir.c_str(), NULL) != 0) // enhance for EAB's??
 #elif defined(__DOS__)
     const wxChar *dirname = dir.c_str();
-  #if defined(__DJGPP__)
+  #if defined(__WATCOMC__)
+    (void)perm;
+    if ( wxMkDir(wxFNSTRINGCAST wxFNCONV(dirname)) != 0 )
+  #elif defined(__DJGPP__)
     if ( mkdir(wxFNCONV(dirname), perm) != 0 )
   #else
     #error "Unsupported DOS compiler!"
@@ -1207,7 +1259,9 @@ bool wxRmdir(const wxString& dir, int WXUNUSED(flags))
 #if defined(__VMS__)
     return false; //to be changed since rmdir exists in VMS7.x
 #else
-  #if defined(__WXWINCE__)
+  #if defined(__OS2__)
+    if ( ::DosDeleteDir(dir.c_str()) != 0 )
+  #elif defined(__WXWINCE__)
     if ( RemoveDirectory(dir.fn_str()) == 0 )
   #else
     if ( wxRmDir(dir.fn_str()) != 0 )
@@ -1262,7 +1316,7 @@ bool wxGetTempFileName(const wxString& prefix, wxString& buf)
 
 // Get first file name matching given wild card.
 
-static wxScopedPtr<wxDir> gs_dir;
+static wxDir *gs_dir = NULL;
 static wxString gs_dirPath;
 
 wxString wxFindFirstFile(const wxString& spec, int flags)
@@ -1273,7 +1327,8 @@ wxString wxFindFirstFile(const wxString& spec, int flags)
     if ( !wxEndsWithPathSeparator(gs_dirPath ) )
         gs_dirPath << wxFILE_SEP_PATH;
 
-    gs_dir.reset(new wxDir(gs_dirPath));
+    delete gs_dir; // can be NULL, this is ok
+    gs_dir = new wxDir(gs_dirPath);
 
     if ( !gs_dir->IsOpened() )
     {
@@ -1292,7 +1347,10 @@ wxString wxFindFirstFile(const wxString& spec, int flags)
     wxString result;
     gs_dir->GetFirst(&result, wxFileNameFromPath(spec), dirFlags);
     if ( result.empty() )
+    {
+        wxDELETE(gs_dir);
         return result;
+    }
 
     return gs_dirPath + result;
 }
@@ -1303,7 +1361,10 @@ wxString wxFindNextFile()
 
     wxString result;
     if ( !gs_dir->GetNext(&result) || result.empty() )
+    {
+        wxDELETE(gs_dir);
         return result;
+    }
 
     return gs_dirPath + result;
 }
@@ -1337,12 +1398,16 @@ wxChar *wxDoGetCwd(wxChar *buf, int sz)
 #else // wxUSE_UNICODE
     bool needsANSI = true;
 
-    #if !defined(HAVE_WGETCWD)
+    #if !defined(HAVE_WGETCWD) || wxUSE_UNICODE_MSLU
         char cbuf[_MAXPATHLEN];
     #endif
 
     #ifdef HAVE_WGETCWD
+        #if wxUSE_UNICODE_MSLU
+            if ( wxGetOsVersion() != wxOS_WINDOWS_9X )
+        #else
             char *cbuf = NULL; // never really used because needsANSI will always be false
+        #endif
             {
                 ok = _wgetcwd(buf, sz) != NULL;
                 needsANSI = false;
@@ -1354,7 +1419,25 @@ wxChar *wxDoGetCwd(wxChar *buf, int sz)
     {
     #if defined(_MSC_VER) || defined(__MINGW32__)
         ok = _getcwd(cbuf, sz) != NULL;
-    #else // !Win32/VC++ !Mac
+    #elif defined(__OS2__)
+        APIRET rc;
+        ULONG ulDriveNum = 0;
+        ULONG ulDriveMap = 0;
+        rc = ::DosQueryCurrentDisk(&ulDriveNum, &ulDriveMap);
+        ok = rc == 0;
+        if (ok)
+        {
+            sz -= 3;
+            rc = ::DosQueryCurrentDir( 0 // current drive
+                                      ,(PBYTE)cbuf + 3
+                                      ,(PULONG)&sz
+                                     );
+            cbuf[0] = char('A' + (ulDriveNum - 1));
+            cbuf[1] = ':';
+            cbuf[2] = '\\';
+            ok = rc == 0;
+        }
+    #else // !Win32/VC++ !Mac !OS2
         ok = getcwd(cbuf, sz) != NULL;
     #endif // platform
 
@@ -1419,6 +1502,13 @@ wxChar *wxDoGetCwd(wxChar *buf, int sz)
     // __WXWINCE__
 }
 
+#if WXWIN_COMPATIBILITY_2_6
+wxChar *wxGetWorkingDirectory(wxChar *buf, int sz)
+{
+    return wxDoGetCwd(buf,sz);
+}
+#endif // WXWIN_COMPATIBILITY_2_6
+
 wxString wxGetCwd()
 {
     wxString str;
@@ -1429,7 +1519,17 @@ wxString wxGetCwd()
 bool wxSetWorkingDirectory(const wxString& d)
 {
     bool success = false;
-#if defined(__UNIX__) || defined(__WXMAC__) || defined(__DOS__)
+#if defined(__OS2__)
+    if (d[1] == ':')
+    {
+        ::DosSetDefaultDisk(wxToupper(d[0]) - wxT('A') + 1);
+    // do not call DosSetCurrentDir when just changing drive,
+    // since it requires e.g. "d:." instead of "d:"!
+    if (d.length() == 2)
+        return true;
+    }
+    success = (::DosSetCurrentDir(d.c_str()) == 0);
+#elif defined(__UNIX__) || defined(__WXMAC__) || defined(__DOS__)
     success = (chdir(wxFNSTRINGCAST d.fn_str()) == 0);
 #elif defined(__WINDOWS__)
 
@@ -1664,7 +1764,7 @@ int WXDLLIMPEXP_BASE wxParseCommonDialogsFilter(const wxString& filterStr,
     return filters.GetCount();
 }
 
-#if defined(__WINDOWS__) && !defined(__UNIX__)
+#if defined(__WINDOWS__) && !(defined(__UNIX__) || defined(__OS2__))
 static bool wxCheckWin32Permission(const wxString& path, DWORD access)
 {
     // quoting the MSDN: "To obtain a handle to a directory, call the
@@ -1707,7 +1807,7 @@ static bool wxCheckWin32Permission(const wxString& path, DWORD access)
 
 bool wxIsWritable(const wxString &path)
 {
-#if defined( __UNIX__ )
+#if defined( __UNIX__ ) || defined(__OS2__)
     // access() will take in count also symbolic links
     return wxAccess(path.c_str(), W_OK) == 0;
 #elif defined( __WINDOWS__ )
@@ -1721,7 +1821,7 @@ bool wxIsWritable(const wxString &path)
 
 bool wxIsReadable(const wxString &path)
 {
-#if defined( __UNIX__ )
+#if defined( __UNIX__ ) || defined(__OS2__)
     // access() will take in count also symbolic links
     return wxAccess(path.c_str(), R_OK) == 0;
 #elif defined( __WINDOWS__ )
@@ -1735,7 +1835,7 @@ bool wxIsReadable(const wxString &path)
 
 bool wxIsExecutable(const wxString &path)
 {
-#if defined( __UNIX__ )
+#if defined( __UNIX__ ) || defined(__OS2__)
     // access() will take in count also symbolic links
     return wxAccess(path.c_str(), X_OK) == 0;
 #elif defined( __WINDOWS__ )
@@ -1808,10 +1908,12 @@ wxFileKind wxGetFileKind(int fd)
 
 wxFileKind wxGetFileKind(FILE *fp)
 {
-#if defined(wxFILEKIND_STUB)
+    // Note: The watcom rtl dll doesn't have fileno (the static lib does).
+    //       Should be fixed in version 1.4.
+#if defined(wxFILEKIND_STUB) || wxONLY_WATCOM_EARLIER_THAN(1,4)
     (void)fp;
     return wxFILE_KIND_DISK;
-#elif defined(__WINDOWS__) && !defined(__CYGWIN__) && !defined(__WINE__)
+#elif defined(__WINDOWS__) && !defined(__CYGWIN__) && !defined(__WATCOMC__) && !defined(__WINE__)
     return fp ? wxGetFileKind(_fileno(fp)) : wxFILE_KIND_UNKNOWN;
 #else
     return fp ? wxGetFileKind(fileno(fp)) : wxFILE_KIND_UNKNOWN;
